Separate update requests from the TelegramClient
